@@ -10,19 +10,19 @@
 	}
   ],
   "require": {
-	"php": ">=7.1",
-	"fapi-cz/http-client": "^0.1"
+	"php": "^8.1",
+	"fapi-cz/http-client": "^1.0"
   },
   "require-dev": {
-	"nette/tester": ">=2.3",
-	"nette/di": "^3.0",
-	"nette/utils": "^3.1",
-	"guzzlehttp/guzzle": "^6.3 | ^7.0",
-	"tracy/tracy": ">=2.7",
-	"phpstan/phpstan-nette": "^0.12.17",
-	"phpstan/phpstan": "^0.12.85",
-    "orisai/coding-standard": "^1.1",
-    "phpstan/phpstan-strict-rules": "^0.12.9"
+	"nette/tester": "^2.5",
+	"nette/di": "^3.1",
+	"nette/utils": "^4.0",
+	"guzzlehttp/guzzle": "6.3 | ^7.0",
+	"tracy/tracy": "^2.10",
+	"phpstan/phpstan-nette": "^1.2",
+	"phpstan/phpstan": "^1.10",
+    "orisai/coding-standard": "^3.10",
+    "phpstan/phpstan-strict-rules": "^1.5"
   },
   "suggest": {
 	"nette/di": "to use class FapiClientExtension"
@@ -31,5 +31,10 @@
 	"psr-0": {
 	  "Fapi\\FapiClient\\": "src/"
 	}
+  },
+  "config": {
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
   }
 }
