@@ -7,18 +7,12 @@ use Fapi\FapiClient\Rest\FapiRestClient;
 final class ExchangeRates
 {
 
-	/** @var FapiRestClient */
-	private $client;
+	private string $path;
 
-	/** @var string */
-	private $path;
+	private string $resources;
 
-	/** @var string */
-	private $resources;
-
-	public function __construct(FapiRestClient $client)
+	public function __construct(private FapiRestClient $client)
 	{
-		$this->client = $client;
 		$this->path = '/exchange-rates';
 		$this->resources = 'exchange_rates';
 	}
