@@ -7,15 +7,10 @@ use Fapi\FapiClient\Rest\FapiRestClient;
 final class Statistics
 {
 
-	/** @var FapiRestClient */
-	private $client;
+	private string $path;
 
-	/** @var string */
-	private $path;
-
-	public function __construct(FapiRestClient $client)
+	public function __construct(private FapiRestClient $client)
 	{
-		$this->client = $client;
 		$this->path = '/statistics';
 	}
 
