@@ -7,17 +7,15 @@ use Fapi\FapiClient\Rest\FapiRestClient;
 trait Find
 {
 
-	/** @var FapiRestClient */
-	private $client;
+	private FapiRestClient $client;
 
-	/** @var string */
-	private $path;
+	private string $path;
 
 	/**
 	 * @param array<mixed> $parameters
 	 * @return array<mixed>|null
 	 */
-	public function find(int $id, array $parameters = []): ?array
+	public function find(int $id, array $parameters = []): array|null
 	{
 		return $this->client->getResource($this->path, $id, $parameters);
 	}
